+++ db/migrate/20260223214109_add_thinking_tokens_to_ruby_llm_monitoring_events.rb
@@ -1,0 +1,7 @@
+class AddThinkingTokensToRubyLLMMonitoringEvents < ActiveRecord::Migration[7.2]
+  include RubyLLM::Monitoring::MigrationHelpers
+
+  def change
+    add_column :ruby_llm_monitoring_events, :thinking_tokens, :virtual, type: :integer, as: json_extract("thinking_tokens", as: :integer), stored: true
+  end
+end